(change_dir_group): Remove cast of savedir arg.

+++ src/chgrp.c
@@ -253,7 +253,7 @@ change_dir_group (const char *dir, gid_t group, const struct stat *statp)
   int errors = 0;
 
   errno = 0;
-  name_space = savedir (dir, (unsigned int) statp->st_size);
+  name_space = savedir (dir, statp->st_size);
   if (name_space == NULL)
     {
       if (errno)
